Catch KeyboardInterupt in download_file
When interupting a download with something like control c the program will quit but won't clean it self up and will leave an invalid file. There is a handler for Exception, but this wont catch interupts. If you try to rerun the program it won't try to redownload it the next time.