Modified code adding warning of possible old Python SSL/TLS certificates to reduce the number of issues raised during SDK installation. The OSX message is [SSL: CERTIFICATE_VERIFY_FAILED] while trying to download files. While, the Linux Python message is less descriptive. Error downloading URL 'https://xxxx.xxx.xxx': <urlopen error unknown url type: https>.
This hopefully reduces the stream of common issues raised during installation as a result of the change of Python TLS certs to 2048 or greater.
This change relates to  #133 #136 #140 #176 #6275 #6548 #6723 #9036 … and possibly more.

Installation can be accomplished by installing the latest Python certificates (pip install certifi), symbolic linking a new python version, and relying on the python shebang line to run $ ./emsdk.py install latest.