Re-enable building master from source for linux 32-bit
Doing this for windows is trickier as we would also need to provide
the python binaries for win32.

Note that we don't provide node binaries in this case because the
node project itself doesn't produce 32-bit builds for linux anymore.
This should still be useable for 32-bit linux users although they
will need to provide their own node (or use the system provided
one).

Fixes: #470