test.py should require that the SDK be activated
It was working previously by accident because the tests were
using the defaults such as $HOME/.emscripten.  However usage of the
emsdk should really only happen in the emsdk environment where
EM_CONFIG dictates the config location (not just using the default).

This change is needed as precursor to making `--embedded` the default
(in this case the config doesn't live in `$HOME/.emscripten`).