Provide "executable" in wasm_cc_binary's DefaultInfo. (#998)
This is needed to allow custom rules and/or transitions on top of wasm_cc_binary (e.g. to workaround #972).

It's debatable whether .wasm is the correct executable target for non-standalone Wasm builds, but we don't have a way to differentiate those in wasm_cc_binary, and considering that this attribute isn't exported now, it shouldn't break anything.