Update EMSDK_OS override
This change modifies the way the EMSDK_OS environment override works.
When I first landed this override it made emsdk behave exactly like the
desired OS.

However what I really want to be able to do is install the files
designed for other OSes, so this change limits it to only effecting the
URLs of the files that get downloaded.