Add home-made emsdk_env.sh (#475)
I was originally going to ship the upstream emsdk_env.sh, but that carries a LOT of dependencies on other files from the emsdk repo, and also the construct_env command doesn't work properly unless the directory layout is very specific, which is a problem for us. Instead, a simple script to add our env vars (which we already use in wasm publishing) to $PATH in the same manner, bypassing the need for emsdk.py entirely.