Unset `PYTHONSAFEPATH` in env.bat and env.sh.
In Python 3.11 and above `PYTHONSAFEPATH` prevents the containing directory of a script from being imported into `sys.path`. This results in `emcc.py` (and other programs) failing with "ModuleNotFoundError: No module named 'tools'". By unsetting `PYTHONSAFEPATH` in env.bat and env.sh, the expected behavior of importing the containing directory of a script to `sys.path` is restored.