Add Actions workflow to automatically create release tags
For commits created by the new create-release.yml workflow, we can add the
corresponding release tag. This only runs for changes to the tag JSON file
(and the workflow itself) and only acts on changes created by the automation
(so that it won't interfere if we want to do things manually).