Pass nodejs binary path as an environment variable.
The current way to derive the binary path relies on a specific name for
the nodejs repository. This blocks migrating to bzlmod, as it generates
repository names based on which module is creating them, resulting in a
different name.

By letting Bazel tell us the correct name and passing it in, we can keep
this working with both old and new name, independent of using bzlmod.