Add support for a new command 'emsdk deactivate' (#1593)
Add support for a new command 'emsdk deactivate' that can be used to
remove a tool from the active list.

Since the big endian cross compile Node.js is added to `NODE_JS_TEST`
instead of `NODE_JS`, there needs to be a way to deactivate it for
regular runs.

This PR adds a `emsdk deactivate` command to enable doing that.